src/handlers/python-test/requirements.txt
selenium==3.141.0
airtable-python-wrapper==0.15.0
boto3==1.17.49
pandas
python-dotenv==0.17.1
openpyxl
tqdm==4.66.1
